fix circular depds
